Adding important notice for Windows User
I have been trying to get this code to work but realised it was not doing anything. After right-clicking the file that I wanted to move and then clicking on properties, the file then disappear and moved to the destinated folder. It's unusal behaviour but I found this trick help me out. Please remove this propose change if it's not appropriate. Thanks!